Refactored io storage
